add aoc.py runner program, adapt days 7 and 8 for it
Welp I did some automation anyway, mainly for generating new days based
on a template, and for running test cases where the input can be stored
inline with the code rather than in a separate file.

This was actually a good learning experience for using importlib and
getattr to dynamically load python modules and run functions and stuff
from them.

It's shebanged for python 3.9 because that's what I'm targeting these at
(assuming there'll be more puzzles that might involve merging dicts).
Arch has 3.9 and in WSL2 I compiled python 3.9 manually.