extra: add a Rust version of Day 15 Part 2
I was lazy and hard-coded the input but wow it's fast. 1.6 seconds
versus Python's ~25.

However the algorithm is different than Python - I only store one most
recent turn in the hashmap rather than a pair, and last/turn aren't
stored in the map until after the next turn is over.