day 20: WIP
A friggin jigsaw puzzle solver, great. Gonna do this one in Rust because
there's too many data types to reason about (tiles and edges and points
and orientations) and I have a feeling that a recursive backtracking
algorithm is going to be slow and I'll want the speed of Rust. Easier
unit tests are a bonus too.

After all, there's 144! * 8^144 = 10^230 possible arrangements of the
board! Who knows how many millions of attempts there will be until we
find a match.

Handy tricks I thought of here:
  * A single edge can be thought of as a 10-bit number, so comparisons
    between edges are just a integer comparison.
  * When pairing tiles, we only care about the edges, so transforming
    the tiles around can be done with operations on four u16s.
  * Reversing a 10-bit number is easy to do with a lookup table (rather
    than a bunch of bitmasks & shifts)

TODO: more tests, literally all of the logic.