day 23: rewrite to be faster
Here we see the introduction of my patented Cursed List - a circular
doubly linked list stored as a flat array with O(1) element lookup!

Using an actual CDLL worked just fine and made sense, but was too slow
for doing 10 million iterations in Part 2. But conveniently the values
are all sequential integers, so we can use that label value as the index
in a flat array.